pybind/mgr/rbd_support: split module.py into several files
(it is becoming too large)
